Remove default terms array, add missing array index
The missing array index just makes sure we're dealing with the right array.

I removed the default terms array out of concern that someone might not pass that array in by mistake and remove all the terms on a post. Now you have to explicitly pass an empty array to do remove a post's terms.